+++ scripts/spells/cloud-of-daggers.js
@@ -7,31 +7,19 @@
 
     Using a Higher-Level Spell Slot. The damage increases by 2d4 for each spell slot level above 2.
  */
-const version = "12.4.0";
+const version = "12.4.1";
 const optionName = "Cloud of Daggers";
-const _flagGroup = "fvtt-trazzm-homebrew-5e";
-const flagName = "cloud-of-daggers-flag";
-const animationFile = "modules/fvtt-trazzm-homebrew-5e/assets/animations/CloudOfDaggers_Kunai_01_Light_Yellow_400x400.webm";
 
 try {
     if (rolledActivity?.name === "Cast") {
         if (args[0].macroPass === "preItemRoll") {
             Hooks.once("createMeasuredTemplate", async (template) => {
-                await actor.setFlag(_flagGroup, flagName, {templateUuid: template.uuid});
-                await template.update({'hidden': true});
-
-                new Sequence()
-                    .effect()
-                    .file(animationFile)
-                    .scaleToObject(1.05)
-                    .aboveInterface()
-                    .opacity(0.75)
-                    .mask(template)
-                    .persist(true)
-                    .attachTo(template)
-                    .play();
-
-            // .xray(true)
+                await template.update({
+                    fillColor: 0,
+                    fillAlpha: 0,
+                    alpha: 0,
+                    opacity: 0.1
+                });
             });
 
             Hooks.once("createRegion", async (region) => {