+++ scripts/spells/web.js
@@ -16,7 +16,7 @@
     any creature that starts its turn in the fire.
  */
 const optionName = "Web";
-const version = "12.4.1";
+const version = "12.4.2";
 const _flagGroup = "fvtt-trazzm-homebrew-5e";
 const textureFile = "modules/JB2A_DnD5e/Library/2nd_Level/Web/Web_01_White_Thumb.webp";
 
@@ -25,8 +25,16 @@ try {
         if (args[0].macroPass === "preItemRoll") {
             Hooks.once("createMeasuredTemplate", async (template) => {
                 console.log(template);
-                await template.update({'fillColor': template.borderColor});
-                await template.update({'texture': textureFile});
+                await template.update({
+                    fillColor: 0,
+                    fillAlpha: 0,
+                    alpha: 0,
+                    opacity: 0.1,
+                    texture: textureFile
+                });
+
+                // await template.update({'fillColor': template.borderColor});
+                // await template.update({'texture': textureFile});
             });
 
             Hooks.once("createRegion", async (region) => {